Added an event handler for the start-game event
When the start-game event is generated by the client, the server should respond to each player with the role they have been allocated. The roles are randomly allocated to each player and broadcasted to them. The server broadcasts the roles along with a game-start event.

The number of available roles of each type depends on the number of players playing the game.