Updated the config file to store the role allocation logic values
The number of roles available of each type depends on the number of players in the game. To ensure no magic numbers are introduced in the code, the values used to calculate the number of roles available for each type have been abstracted out into the config file.

These values can be changed later if the game rules change.